Code review, mostly AbstractShapeIndexFieldData
