+++ x-pack/plugin/spatial/src/main/java/org/elasticsearch/xpack/spatial/index/fielddata/plain/LatLonShapeIndexFieldData.java
@@ -7,22 +7,15 @@
 
 package org.elasticsearch.xpack.spatial.index.fielddata.plain;
 
-import org.apache.lucene.index.DocValuesType;
 import org.apache.lucene.index.FieldInfo;
 import org.apache.lucene.index.LeafReader;
 import org.apache.lucene.index.LeafReaderContext;
-import org.elasticsearch.common.util.BigArrays;
 import org.elasticsearch.script.field.ToScriptFieldFactory;
-import org.elasticsearch.search.DocValueFormat;
-import org.elasticsearch.search.MultiValueMode;
 import org.elasticsearch.search.aggregations.support.ValuesSourceType;
-import org.elasticsearch.search.sort.BucketedSort;
-import org.elasticsearch.search.sort.SortOrder;
-import org.elasticsearch.xpack.spatial.index.fielddata.IndexShapeFieldData;
 import org.elasticsearch.xpack.spatial.index.fielddata.LeafShapeFieldData;
 import org.elasticsearch.xpack.spatial.index.fielddata.ShapeValues;
 
-public class LatLonShapeIndexFieldData extends AbstractShapeIndexFieldData implements IndexShapeFieldData {
+public class LatLonShapeIndexFieldData extends AbstractShapeIndexFieldData {
     public LatLonShapeIndexFieldData(
         String fieldName,
         ValuesSourceType valuesSourceType,
@@ -36,43 +29,13 @@ public class LatLonShapeIndexFieldData extends AbstractShapeIndexFieldData imple
         LeafReader reader = context.reader();
         FieldInfo info = reader.getFieldInfos().fieldInfo(fieldName);
         if (info != null) {
-            checkCompatible(info);
+            checkCompatible(info, "geo_shape");
         }
         return new LatLonShapeDVAtomicShapeFieldData(reader, fieldName, toScriptFieldFactory);
     }
 
     @Override
-    public LeafShapeFieldData loadDirect(LeafReaderContext context) {
-        return load(context);
-    }
-
-    @Override
-    public BucketedSort newBucketedSort(
-        BigArrays bigArrays,
-        Object missingValue,
-        MultiValueMode sortMode,
-        XFieldComparatorSource.Nested nested,
-        SortOrder sortOrder,
-        DocValueFormat format,
-        int bucketSize,
-        BucketedSort.ExtraData extra
-    ) {
-        throw new IllegalArgumentException("can't sort on geo_shape field without using specific sorting feature, like geo_distance");
-    }
-
-    /** helper: checks a fieldinfo and throws exception if its definitely not a LatLonDocValuesField */
-    static void checkCompatible(FieldInfo fieldInfo) {
-        // dv properties could be "unset", if you e.g. used only StoredField with this same name in the segment.
-        if (fieldInfo.getDocValuesType() != DocValuesType.NONE && fieldInfo.getDocValuesType() != DocValuesType.BINARY) {
-            throw new IllegalArgumentException(
-                "field=\""
-                    + fieldInfo.name
-                    + "\" was indexed with docValuesType="
-                    + fieldInfo.getDocValuesType()
-                    + " but this type has docValuesType="
-                    + DocValuesType.BINARY
-                    + ", is the field really a geo-shape field?"
-            );
-        }
+    protected IllegalArgumentException sortException() {
+        return new IllegalArgumentException("can't sort on geo_shape field without using specific sorting feature, like geo_distance");
     }
 }
